+++ metro_benchmark_msgs/CHANGELOG.rst
@@ -1,5 +1,5 @@
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-Changelog for package compute_benchmark_msgs
+Changelog for package metro_benchmark_msgs
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
 0.1.0 (2023-08-29)
