+++ metro_benchmark_msgs/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>compute_benchmark_msgs</name>
+  <name>metro_benchmark_msgs</name>
   <version>0.1.0</version>
   <description>ROS interfaces for recording compute time and other related benchmarking concepts</description>
   <maintainer email="davidvlu@gmail.com">David V. Lu!!</maintainer>
