+++ metro_benchmark_pub/CHANGELOG.rst
@@ -1,5 +1,5 @@
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-Changelog for package compute_benchmarking
+Changelog for package metro_benchmark_pub
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
 0.1.0 (2023-08-29)
