+++ metro_benchmark_pub/metro_benchmark_pub/__init__.py
@@ -1,4 +1,4 @@
-from compute_benchmark_msgs.msg import ComputeTime
+from metro_benchmark_msgs.msg import ComputeTime
 import collections
 import math
 import time
