+++ metro_benchmark_pub/package.xml
@@ -1,15 +1,15 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>compute_benchmarking</name>
+  <name>metro_benchmark_pub</name>
   <version>0.1.0</version>
-  <description>Utilities for publishing / processing compute_benchmark_msgs</description>
+  <description>Utilities for publishing / processing metro_benchmark_msgs</description>
   <maintainer email="davidvlu@gmail.com">David V. Lu!!</maintainer>
   <license>BSD 3-clause</license>
 
   <buildtool_depend>ament_cmake</buildtool_depend>
   <buildtool_depend>ament_cmake_python</buildtool_depend>
-  <depend>compute_benchmark_msgs</depend>
+  <depend>metro_benchmark_msgs</depend>
   <depend>rclcpp</depend>
   <test_depend>ament_cmake_gtest</test_depend>
   <test_depend>ament_cmake_pytest</test_depend>
